refactored databaseFunctions -> orderToDB for better code clarity
